+++ src/services/MapService.php
@@ -314,8 +314,12 @@ class MapService extends Component
 		// Normalize location
 		$location = GeoService::normalizeLocation($location, $country);
 
+		// If we don't have a location, reduce the search radius to 0
 		if ($location === null)
-			return false;
+		{
+			$location = ['lat' => 0, 'lng' => 0];
+			$radius = 0;
+		}
 
 		$lat = $location['lat'];
 		$lng = $location['lng'];